test(rpc): address review on creation-deposit tests
- Move the tests to a dedicated state_tests.rs file to keep state.rs
  from growing further.
- Build the actor amounts through the TokenAmount shim (from_whole)
  rather than naming actor/fvm versions directly, and add anyhow context
  to the state reads in compute_initial_pledge_for_power.
- Exercise network version 27 through calibnet's real activation epoch
  and a promoted head, instead of overriding the genesis network
  version; keep the pre-v27 case on mainnet at genesis.
- Assert fixed, deterministic pledge values, with the mid-ramp pledge at
  roughly half the fully-ramped one so the ramp parameters are covered.
